# tests/contract/api-gateway-openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "WorkflowApprovalDecisionRequest": {
        "properties": {
          "approver_id": {
            "title": "Approver Id",
            "type": "string"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "approver_id"
        ],
        "title": "WorkflowApprovalDecisionRequest",
        "type": "object"
      },
      "WorkflowApprovalDetailResponse": {
        "properties": {
          "approval_id": {
            "title": "Approval Id",
            "type": "string"
          },
          "approver_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approver Id"
          },
          "audit_trail": {
            "items": {
              "type": "object"
            },
            "title": "Audit Trail",
            "type": "array"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "run_id",
          "step_id",
          "tenant_id",
          "status",
          "created_at",
          "updated_at",
          "decision",
          "approver_id",
          "comments",
          "metadata",
          "audit_trail"
        ],
        "title": "WorkflowApprovalDetailResponse",
        "type": "object"
      },
      "WorkflowApprovalResponse": {
        "properties": {
          "approval_id": {
            "title": "Approval Id",
            "type": "string"
          },
          "approver_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approver Id"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "run_id",
          "step_id",
          "tenant_id",
          "status",
          "created_at",
          "updated_at",
          "decision",
          "approver_id",
          "comments",
          "metadata"
        ],
        "title": "WorkflowApprovalResponse",
        "type": "object"
      },
      "WorkflowDefinitionCreateRequest": {
        "properties": {
          "definition": {
            "title": "Definition",
            "type": "object"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "definition"
        ],
        "title": "WorkflowDefinitionCreateRequest",
        "type": "object"
      },
      "WorkflowDefinitionResponse": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "name",
          "version",
          "owner",
          "description"
        ],
        "title": "WorkflowDefinitionResponse",
        "type": "object"
      },
...
      "WorkflowRunResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step Id"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "workflow_id",
          "tenant_id",
          "status",
          "current_step_id",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowRunResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-native Project Portfolio Management platform with 25 specialized agents",
    "title": "Multi-Agent PPM Platform",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workflows/approvals": {
      "get": {
        "operationId": "list_approvals_api_v1_workflows_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "approver_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Approver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowApprovalResponse"
                  },
                  "title": "Response List Approvals Api V1 Workflows Approvals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Approvals",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/approvals/{approval_id}": {
      "get": {
        "operationId": "get_approval_api_v1_workflows_approvals__approval_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowApprovalDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Approval",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/approvals/{approval_id}/decision": {
      "post": {
        "operationId": "decide_approval_api_v1_workflows_approvals__approval_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Approval",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/definitions": {
      "get": {
        "operationId": "list_definitions_api_v1_workflows_definitions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowDefinitionResponse"
                  },
                  "title": "Response List Definitions Api V1 Workflows Definitions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Definitions",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "create_definition_api_v1_workflows_definitions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDefinitionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDefinitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Definition",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/v1/workflows/start": {
      "post": {
        "operationId": "start_workflow_api_v1_workflows_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}